pdf-service/requirements.txt

sh
flask
flask_cors
pdfjinja
requests
reportlab
pypdf2
pillow
fdfgen
raven[flask]
